refactor: move DocumentSummary DTO to KnowledgeBase Contracts namespace
Extracts DocumentSummary class from GetDocumentsRequest.cs into a dedicated
Contracts/DocumentSummary.cs file, updating all consumers to reference the
new canonical location.
