+++ backend/src/Anela.Heblo.Application/Features/KnowledgeBase/UseCases/GetDocuments/GetDocumentsHandler.cs
@@ -1,3 +1,4 @@
+using Anela.Heblo.Application.Features.KnowledgeBase.Contracts;
 using Anela.Heblo.Domain.Features.KnowledgeBase;
 using MediatR;
