+++ backend/src/Anela.Heblo.Application/Features/KnowledgeBase/UseCases/UploadDocument/UploadDocumentResponse.cs
@@ -1,4 +1,4 @@
-using Anela.Heblo.Application.Features.KnowledgeBase.UseCases.GetDocuments;
+using Anela.Heblo.Application.Features.KnowledgeBase.Contracts;
 using Anela.Heblo.Application.Shared;
 
 namespace Anela.Heblo.Application.Features.KnowledgeBase.UseCases.UploadDocument;
